Мелкие правки для соответствия макету
