+++ FakeNFT/DesignSystem/ExtensionUINavigationBar.swift
@@ -1,0 +1,20 @@
+//
+//  ExtensionUINavigationBar.swift
+//  FakeNFT
+//
+//  Created by Aleksey Yakushev on 30.10.2023.
+//
+
+import UIKit
+
+extension UINavigationBar {
+    open override func didAddSubview(_ subview: UIView) {
+        super.didAddSubview(subview)
+        
+        let subviewClass = String(describing: type(of: subview))
+        print(subviewClass)
+        if subviewClass.lowercased().localizedStandardContains("background") {
+            subview.isHidden = true
+        }
+    }
+}
