+++ FakeNFT/MainTabBarViewController.swift
@@ -4,17 +4,6 @@ import UIKit
 
 final class ProfileViewController: UIViewController { }
 
-extension UITabBar {
-    open override func didAddSubview(_ subview: UIView) {
-        super.didAddSubview(subview)
-        
-        let subviewClass = String(describing: type(of: subview))
-        if subviewClass == "UIBarBackgroundShadowContentImageView" {
-            subview.isHidden = true
-        }
-    }
-}
-
 final class MainTabBarViewController: UITabBarController {
     let statisticVC = UINavigationController(rootViewController: StatisticsViewController())
     let cartVC = UINavigationController(rootViewController: CartViewController())
